+++ policy_set_integration_test.go
@@ -159,6 +159,7 @@ func TestPolicySetsCreate(t *testing.T) {
 		Beta:       Bool(false),
 	}
 	sv, err := client.Admin.SentinelVersions.Create(ctx, opts)
+
 	defer func() {
 		err := client.Admin.SentinelVersions.Delete(ctx, sv.ID)
 		require.NoError(t, err)
@@ -414,6 +415,7 @@ func TestPolicySetsCreate(t *testing.T) {
 				OAuthTokenID:      String(oc.ID),
 				IngressSubmodules: Bool(true),
 			},
+			PolicyUpdatePatterns: []*string{String("**"), String("*.sentinel")},
 		}
 
 		ps, err := client.PolicySets.Create(ctx, orgTest.Name, options)
@@ -435,6 +437,9 @@ func TestPolicySetsCreate(t *testing.T) {
 		assert.Equal(t, ps.VCSRepo.RepositoryHTTPURL, fmt.Sprintf("https://github.com/%s", githubIdentifier))
 		assert.Equal(t, ps.VCSRepo.ServiceProvider, string(ServiceProviderGithub))
 		assert.Regexp(t, fmt.Sprintf("^%s/webhooks/vcs/[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}$", regexp.QuoteMeta(DefaultConfig().Address)), ps.VCSRepo.WebhookURL)
+
+		assert.Equal(t, len(ps.PolicyUpdatePatterns), 2)
+		assert.Contains(t, ps.PolicyUpdatePatterns, "*.sentinel")
 	})
 
 	t.Run("with vcs policy updated", func(t *testing.T) {
@@ -460,6 +465,7 @@ func TestPolicySetsCreate(t *testing.T) {
 				OAuthTokenID:      String(oc.ID),
 				IngressSubmodules: Bool(false),
 			},
+			PolicyUpdatePatterns: []*string{String("**"), String("*.sentinel")},
 		}
 
 		ps, err := client.PolicySets.Update(ctx, vcsPolicyID, options)
@@ -477,6 +483,9 @@ func TestPolicySetsCreate(t *testing.T) {
 		assert.Equal(t, ps.VCSRepo.RepositoryHTTPURL, fmt.Sprintf("https://github.com/%s", githubIdentifier))
 		assert.Equal(t, ps.VCSRepo.ServiceProvider, string(ServiceProviderGithub))
 		assert.Regexp(t, fmt.Sprintf("^%s/webhooks/vcs/[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}$", regexp.QuoteMeta(DefaultConfig().Address)), ps.VCSRepo.WebhookURL)
+
+		assert.Equal(t, len(ps.PolicyUpdatePatterns), 1)
+		assert.Contains(t, ps.PolicyUpdatePatterns, "*.sentinel")
 	})
 
 	t.Run("without a name provided", func(t *testing.T) {